backend/requirements.txt
# Django core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2

# Async / ASGI
channels==4.1.0
channels-redis==4.2.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.4

# OKX
python-okx==0.3.9

# Telegram
python-telegram-bot==21.3

# Utilities
httpx==0.27.0
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
